# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/appraisals:
    get:
      tags:
        - Appraisals
      summary: List Appraisals
      description: Get a paginated list of Appraisals with optional filtering and sorting
      operationId: listAppraisals
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppraisalList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Appraisals
      summary: Create Appraisal
      description: Create a new Appraisal
      operationId: createAppraisal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppraisalRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appraisal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/appraisal_cycles:
    get:
      tags:
        - AppraisalCycles
      summary: List AppraisalCycles
      description: Get a paginated list of AppraisalCycles with optional filtering and sorting
      operationId: listAppraisalCycles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppraisalCycleList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - AppraisalCycles
      summary: Create AppraisalCycle
      description: Create a new AppraisalCycle
      operationId: createAppraisalCycle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppraisalCycleRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppraisalCycle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Appraisal:
      type: object
      required:
        - id
        - employee_id
        - cycle_id
        - reviewer_id
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        cycle_id:
          type: string
          format: uuid
        reviewer_id:
          type: string
          format: uuid
        status:
          type: string
        overall_rating:
          nullable: true
          type: string
        submitted_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateAppraisalRequest:
      type: object
      required:
        - employee_id
        - cycle_id
        - reviewer_id
        - status
        - metadata
      properties:
        employee_id:
          type: string
          format: uuid
        cycle_id:
          type: string
          format: uuid
        reviewer_id:
          type: string
          format: uuid
        status:
          type: string
        overall_rating:
          nullable: true
          type: string
        submitted_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    AppraisalList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Appraisal'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    AppraisalCycle:
      type: object
      required:
        - id
        - name
        - period_start
        - period_end
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 180
        cycle_type:
          nullable: true
          type: string
          maxLength: 40
        period_start:
          type: string
          format: date
        period_end:
          type: string
          format: date
        status:
          type: string
        metadata:
          type: object
...
    CreateAppraisalCycleRequest:
      type: object
      required:
        - name
        - period_start
        - period_end
        - status
        - metadata
      properties:
        name:
          type: string
          maxLength: 180
        cycle_type:
          nullable: true
          type: string
          maxLength: 40
        period_start:
          type: string
          format: date
        period_end:
          type: string
          format: date
        status:
          type: string
        metadata:
          type: object
...
    AppraisalCycleList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AppraisalCycle'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []